Admin/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
loguru==0.7.2
sqlparse==0.4.4
